[refactor] ScriptCommandInvoker, ShellCommandInvoker 통합 (-> CommandInvoker)
