+++ shell/src/main/java/com/samsung/TestShellApplication.java
@@ -2,12 +2,11 @@ package com.samsung;
 
 import com.samsung.file.FileManager;
 import com.samsung.file.JarExecutor;
-import com.samsung.testscript.ScriptCommandInvoker;
 import com.samsung.testscript.ScriptManager;
 import com.samsung.testscript.command.TestScript1Command;
 import com.samsung.testscript.command.TestScript2Command;
 import com.samsung.testscript.command.TestScript3Command;
-import com.samsung.testshell.ShellCommandInvoker;
+import com.samsung.testshell.CommandInvoker;
 import com.samsung.testshell.TestShellManager;
 import com.samsung.testshell.commands.*;
 
@@ -15,12 +14,11 @@ import java.util.Scanner;
 
 public class TestShellApplication {
     public static void main(String[] args) {
-        ShellCommandInvoker shellCommandInvoker = new ShellCommandInvoker();
-        ScriptCommandInvoker scriptCommandInvoker = new ScriptCommandInvoker();
+        CommandInvoker commandInvoker = new CommandInvoker();
 
         // 명령어 등록
-        initShellCommand(shellCommandInvoker);
-        initScriptCommand(scriptCommandInvoker);
+        initShellCommand(commandInvoker);
+        initScriptCommand(commandInvoker);
 
         // 사용자 입력 받기
         System.out.println("testshell 명령어: write, read, exit, help, fullwrite, fullread");
@@ -43,9 +41,9 @@ public class TestShellApplication {
                 }
 
                 if (commandName.contains("_")) {
-                    scriptCommandInvoker.execute(cmdArgs);
+                    commandInvoker.execute(cmdArgs);
                 } else {
-                    shellCommandInvoker.execute(cmdArgs);
+                    commandInvoker.execute(cmdArgs);
                 }
             } catch (Exception e) {
                 System.out.println(e.getMessage());
@@ -54,21 +52,21 @@ public class TestShellApplication {
 
     }
 
-    private static void initShellCommand(ShellCommandInvoker shellCommandInvoker) {
+    private static void initShellCommand(CommandInvoker commandInvoker) {
         TestShellManager testShellManager = new TestShellManager(new JarExecutor(), new FileManager());
-        shellCommandInvoker.register("write", new WriteCommand(testShellManager));
-        shellCommandInvoker.register("read", new ReadCommand(testShellManager));
-        shellCommandInvoker.register("exit", new ExitCommand(testShellManager));
-        shellCommandInvoker.register("help", new HelpCommand(testShellManager));
-        shellCommandInvoker.register("fullwrite", new FullWriteCommand(testShellManager));
-        shellCommandInvoker.register("fullread", new FullReadCommand(testShellManager));
+        commandInvoker.register("write", new WriteCommand(testShellManager));
+        commandInvoker.register("read", new ReadCommand(testShellManager));
+        commandInvoker.register("exit", new ExitCommand(testShellManager));
+        commandInvoker.register("help", new HelpCommand(testShellManager));
+        commandInvoker.register("fullwrite", new FullWriteCommand(testShellManager));
+        commandInvoker.register("fullread", new FullReadCommand(testShellManager));
     }
 
-    private static void initScriptCommand(ScriptCommandInvoker scriptCommandInvoker) {
+    private static void initScriptCommand(CommandInvoker commandInvoker) {
         ScriptManager scriptManager = new ScriptManager(new FileManager(), new JarExecutor());
-        scriptCommandInvoker.register("1_FullWriteAndReadCompare", new TestScript1Command(scriptManager));
-        scriptCommandInvoker.register("2_PartialLBAWrite", new TestScript2Command(scriptManager));
-        scriptCommandInvoker.register("3_WriteReadAging", new TestScript3Command(scriptManager));
+        commandInvoker.register("1_FullWriteAndReadCompare", new TestScript1Command(scriptManager));
+        commandInvoker.register("2_PartialLBAWrite", new TestScript2Command(scriptManager));
+        commandInvoker.register("3_WriteReadAging", new TestScript3Command(scriptManager));
     }
 
     private static String getInput() {
